Make test pass for decoding json in to a list of games
Removing the first row of the table (which has headers and cannot be converted) was very taxing on my brain. It still doesn't seem like it is done very well. I think there is still scope to remove the gameFromStrings function and replace it with Game, but haven't got it figured out yet