Add routing (work in progress)
The routing now mostly works, in that you can type a url
(for example http://localhost:3000/league/Regional%20Div%201) in to the browser
and it loads the correct page

When you click on the elements in the app, they don't update the url with the current
location though.

There is no end to end test for this, as elm-spec doesn't use a browser, so I just have
to live with that.

I need to add a test for parseLocation.

And there are some tidy up type jobs
- State and Route coexist and basically do the same thing
- Need to unencode leagueTitle from location
- Need to have a 404 page
- need to be consistent about union type names (for example should I use
LeagueListRoute, or is LeagueList, in the Route module, enough)