Update the url when navigating within the app
I initially tried this by using html link elements, but this causes a full page reload and a lot of Jank, so instead I set the url in the browser / history.

When running elm-spec tests, changes to the location do not trigger OnLocation messages (as there is no browser). This is a bit annoying, and means we have to jump through some more hoops to keep the route and the url in sync, and to avoid an infinite loop.