Use RemoteData for LeagueList page
This means that I can show a "loading" message while waiting for the google sheets api to return (which might be a long time on a mobile).

The "loading" message is currently just text, might add some fancy animation or something later.

When refreshing the page, or returning to the page, any existing league display (from a previous fetch from the api) dissapears, to be replaced by the loading message. It would be good to keep the existing league list, as it is unlikely to change very often. RemoteData doesn't really support this, but I could roll my own.

If the api returns quickly, the "loading" message flashes up and dissapears before it is readable. It would be good to make this better. Maybe with a fade in fade out or similar.

I now need to do the same for LeagueTable.