Convert "loading" text to an animation
The animation starts after a delay, to avoid Jank. This is not supported in style-elements so it is done in html / css instead.

Also add more specific / helpful error messages, although there is still some more improvements to make.

Removed the default elm-app css, and move it in to LeagueStyleElements.