Support unplayed games
It is necessary to allow unplayed games to be present in the google spreadsheet. Otherwise, at the start of the season the league table will be blank, whereas we would like it to be the full list of teams that are in the league (but with zero games played).

This commit increases the code duplication problem in LeagueTableFromLeagueGames, which needs fixing.