Sprout function to call google api endpoint for an individual sheet
This is now used in two places, so sharing some code. The methods to call the api are quite a long way apart. I think this was ok when there was one call per page, but now there is some sharing, I think it would be better to keep all the api calls together in a module