Introduce a type to represent a page
The idea is that this will help to reduce duplication, and help
enforce a consistent style.

The HeaderBar will only accept HeaderBarItems, so it won't
be easy to put a HeaderBarButton in the wrong part of the
page, and suchlike. I will also be less likely to make a typo and
mess up a page, as the views will just create the type, and
all the rendering happens in one place, and won't need
changing very often.

It does add some code / overhead, but overall I'm pretty happy
with the plan