Update results fixture view to use new ResultsFixtures type
The unit and end to end tests now pass

There is still some code tidy up to do

The UI doesn't look right yet, and doesn't contain headers
for each date. I need to update the end to end test to require
this

I need to add test for order of results fixtures (descending
date, unscheduled go last)

I can now remove LeagueGames from Model. Not sure if
this is the right thing to do or not. It might be handy later
when I reduce the calls to the Api.

I need to check the existing ResultsFixtures update tests to
see if they still make sense