Remove home grown date formatting function
This is not used any more. I wrote it (instead of using
Date.Extra) intially to avoid bringing in the dependency,
as the Netlify build was timing out. I now don't run the
tests as part of the Netfliy build (and do this on AppVeyor
instead), so this is less of an issue.