Share update code related to api request for single sheet of spreadsheet
This has made things a bit simpler and removed an uneccesary Msg,
and it will make it easier to refactor to only call the Api if the data
isn't already available.

The only possible downside is that is does more calculation than is
strictly necessary, but computers are fast and it is small data and
network speed will definitely be factors of 10 more important.