Store a LeagueTable for each League in the model
At the moment this is just extra code, but it allows the next change,
which is to only call the api if we don't already have the data.

This would have been possible before, but only within the same league,
if you switched leagues you would always have had to call the api. Once
the next change is complete, you will only have to call the api the
first time you view data for a league.