Fix browser navigation / back button
There were two problems with the browser navigation.

One was that the league titles from the sheet weren't being encoded,
so they looked different when I set them in the browser, and when the
browser returned them. This is fixed by encoding / decoding.

The other was that when the browser notified us of a location change,
we added the new location to the history, which doesn't make sense, as
it is already in the history. this led to strange results, mostly
having to click back twice to actually do anything.