Create animation to show fetch failed message
This is the start of a larger piece of work to store the api results in
local storage, for cases when there is no network / response from the
api. Before you would have got a failure screen when there was a problem
calling the api, but if it is retrieved from the cache, then you might
see nothing. I think you need to be notified of the failure, so that
you know that the data is stale and that you might want to try and
refresh it again later