Work with relative urls in development
This requires adding the elmapp.config.js, which create-elm-app uses
to modify webpack on our behalf.

I found that calling the api took minutes initially, even when using the
same url took seconds when used in the browser. Adding the Host and
Accept headers fixed this problem.

This commit also containts some code to try and use the .env file
within the function for the environment variables, but this isn't
working yet. Instead I am just setting the environment variables
manually for the time being