Only record the cypress tests on appveyor
AppVeyor has the cypress key installed as a secure environment variable.

Netlify could have it, but it would have to happen for every site, and
other people apart from me could deploy it, and they wouldn't have the
key.

The same applies for testing locally, if someone else clones the repo,
it is better that they can run the tests immediately.