Upgrade to elm 0.19
There were a lot of changes and it was difficult to commit them sensibly,
so this is a big commit.

The tests now pass, but I haven't tried running the app yet.

The html tests have all been removed for now, as I couldn't find an
html test package for 0.19.

The browser navigation is currently disabled, as there is an annoyance
with a new Key value, that is going to require some rework,