Restore html tests
It turns out that the package has been moved to elm-exporations/test.

My view code now returns a list of Html msg, which is a bit annoying
to test, although actually it always returns a list with one item in it,
so am exposing this function to make the testing easier.