Update usage_scenario after messages with Arne from the tool
the "expose" setting is turned off in our live testing service, as it
would mean that you could expose everything to the outside world. This
is  a security measure.

It is not yet documented I noticed, but we will take of that.
Thanks for the hint.

2. The enviroment variables are not accepted as the contain problematic
characters. Everything else than A-Za-z_- is forbidden as whitelisting
mechanism. Also something that lacks documentation :(

3. However the route you went in the end of passing the variables
through the config should actually work.

The issue I spotted is that you are not explicitely putting the
containers onto a network.

Docker compose does this, but docker-run which we use as underlying
provider does not do it.